изменение: подправил .proto файлы
