+++ app/src/main/kotlin/auth/Authenticator.kt
@@ -1,5 +1,10 @@
 package auth
 
+import grpc.AuthProto.AuthResponse
+import io.grpc.StatusRuntimeException
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
 /**
  * Интерфейс объекта для отправки запросов аутентификации на сервер
  */
@@ -10,4 +15,24 @@ interface Authenticator {
 
     suspend fun register(name: String, login: String, password: String): Result<String>
     suspend fun login(login: String, password: String): Result<String>
+
+    // Вспомогательная функция для выполнения gRPC вызовов с обработкой ошибок
+    suspend fun executeGrpcCall(call: () -> AuthResponse): Result<String> =
+        withContext(Dispatchers.IO) {
+            try {
+                val response = call()
+                when (response.resultCode) {
+                    0 -> Result.success(response.message)
+                    1 -> Result.failure(Authenticator.UserAlreadyExistsException(response.message))
+                    2 -> Result.failure(Authenticator.InvalidCredentialsException(response.message))
+                    else -> Result.failure(Exception(response.message))
+                }
+            } catch (e: StatusRuntimeException) {
+                Result.failure(
+                    Authenticator.ServerConnectionException(
+                        "Failed to connect to the server: server is unavailable"
+                    )
+                )
+            }
+        }
 }
